# fix: resolve query failures in clear-reindex and cross-process scenarios