--- a/src/oboyu/cli/index.py
+++ b/src/oboyu/cli/index.py
@@ -11,7 +11,7 @@ import typer
 from typing_extensions import Annotated
 
 from oboyu.cli.formatters import ProgressGroup, console
-from oboyu.cli.paths import DEFAULT_DB_PATH
+from oboyu.common.paths import DEFAULT_DB_PATH
 from oboyu.indexer.config import IndexerConfig
 from oboyu.indexer.indexer import Indexer