--- a/src/oboyu/common/__init__.py
+++ b/src/oboyu/common/__init__.py
@@ -1,0 +1,5 @@
+"""Common utilities and shared components for Oboyu.
+
+This package contains shared functionality that can be used across
+different modules without creating circular dependencies.
+"""