--- a/tests/integration/__init__.py
+++ b/tests/integration/__init__.py
@@ -1,0 +1,5 @@
+"""Integration tests for Oboyu.
+
+This package contains integration tests that test the complete workflows
+and interactions between different components of the system.
+"""