… src/oboyu/common/paths.py
… tests/indexer/test_indexer.py
… tests/integration/test_cli_clear_index_query.py
… tests/integration/test_debug_clear_issue.py
… tests/integration/test_embedding_investigation.py